Refactor header to component
