+++ assets/ts/app.ts
@@ -2,7 +2,7 @@ import "../scss/app.scss";
 
 import "regenerator-runtime/runtime";
 
-import Header from "./modules/header";
+import Header from "../../templates/components/header/header";
 import LazyLoading from "./modules/lazy-loading";
 import Analytics from "./modules/analytics";
 import CookieSettings from "../../templates/components/cookie-settings/cookie-settings";
