+++ templates/base.html.twig
@@ -80,7 +80,7 @@
 
     </head>
     <body class="header-offset">
-        {% include '/includes/header.html.twig' %}
+        {% include '/components/header/header.html.twig' %}
         <main class="main">
             {% block main %}{% endblock %}
         </main>
