+++ templates/components/header/_header.scss
@@ -1,9 +1,9 @@
-@use "../config/colors" as c;
-@use "../config/decorations" as d;
-@use "../config/z-indexes" as z;
-@use "../config/sizes" as s;
-@use "../config/spacings" as sp;
-@use "../config/breakpoints" as bp;
+@use "assets/scss/config/colors" as c;
+@use "assets/scss/config/decorations" as d;
+@use "assets/scss/config/z-indexes" as z;
+@use "assets/scss/config/sizes" as s;
+@use "assets/scss/config/spacings" as sp;
+@use "assets/scss/config/breakpoints" as bp;
 
 .header {
   background-color: c.$beige;
