+++ templates/components/header/header.html.twig
@@ -1,0 +1,7 @@
+<header class="header" id="js-header">
+    <div class="header__logo">
+        {% include('/components/header/logo/logo.html.twig') %}
+    </div>
+    {% include('/components/header/navigation/navigation.html.twig') %}
+    {{ include('/components/header/language-link/language-link.html.twig') }}
+</header>
