+++ templates/components/header/language-link/_language-link.scss
@@ -1,8 +1,8 @@
-@use "../config/colors" as c;
-@use "../config/breakpoints" as bp;
-@use "../config/sizes" as s;
-@use "../config/spacings" as sp;
-@use "../config/functions" as f;
+@use "assets/scss/config/colors" as c;
+@use "assets/scss/config/breakpoints" as bp;
+@use "assets/scss/config/sizes" as s;
+@use "assets/scss/config/spacings" as sp;
+@use "assets/scss/config/functions" as f;
 
 $flag-size: 16px;
