+++ templates/components/header/logo/_logo.scss
@@ -1,7 +1,7 @@
-@use "../config/breakpoints" as bp;
-@use "../config/spacings" as sp;
-@use "../config/functions" as f;
-@import "../config/silent-classes";
+@use "assets/scss/config/breakpoints" as bp;
+@use "assets/scss/config/spacings" as sp;
+@use "assets/scss/config/functions" as f;
+@import "assets/scss/config/silent-classes";
 
 .logo {
   @extend %text-colored-link;
