+++ templates/components/header/navigation/_navigation.scss
@@ -1,9 +1,9 @@
-@use "../config/breakpoints" as bp;
-@use "../config/spacings" as sp;
-@use "../config/sizes" as s;
-@use "../config/functions" as f;
-@use "../config/colors" as c;
-@import "../config/silent-classes";
+@use "assets/scss/config/breakpoints" as bp;
+@use "assets/scss/config/spacings" as sp;
+@use "assets/scss/config/sizes" as s;
+@use "assets/scss/config/functions" as f;
+@use "assets/scss/config/colors" as c;
+@import "assets/scss/config/silent-classes";
 
 .navigation {
   @extend %smooth-scroll;
