--- templates/includes/header.html.twig
@@ -1,7 +1,0 @@
-<header class="header" id="js-header">
-    <div class="header__logo">
-        {% include('/includes/header/logo.html.twig') %}
-    </div>
-    {% include('/includes/header/navigation.html.twig') %}
-    {{ include('/includes/header/language-switcher.html.twig') }}
-</header>
